Allow subprojects to be published using JitPack
In order for JitPack to publish the subprojects at the expected
coordinates, at least two subprojects must be published.  Otherwise, the
single subproject is published at the coordinates for the repo as a
whole.  The JitPack coordinates for multi-module build artifacts are
described in [1].

Because this repo contains both Gradle and Maven subprojects, a custom
JitPack build script is used to build the individual subprojects.  This
differs from the canonical JitPack multi-module build examples (see [2]
and [3]), which use a single build tool per repo, and thus can use the
build tool's multi-module project support.

[1] https://jitpack.io/docs/BUILDING/#multi-module-projects
[2] https://github.com/jitpack/gradle-modular
[3] https://github.com/jitpack/maven-modular