Use Bazelisk when running the Bazel-based sample
Bazelisk is a command-line tool similar to the Gradle Wrapper in that
it is a commit-able script which downloads a version of Bazel, stores
it in a common location, and then runs that copy of Bazel whenever it
is needed, eliminating the need to have Bazel pre-installed.

The version of Bazelisk included is from commit
https://github.com/philwo/bazelisk/commit/b6727905b0b8c0fe67fb4de1be3c4816640b372d

Currently, version 0.21.0 of Bazel is downloaded and used, according
to the contents of "junit5-jupiter-starter-bazel/.bazelversion".

There is potential to cache the Bazel executable which is downloaded
during CI runs on Travis CI. This is left as a future endeavour.